requirements.txt
lark
pydot
pillow
nbformat
ipython
numpy
aalpy
graphviz
networkx
matplotlib
dash-bootstrap-components
pandas
dash_svg
dash
fastapi
pydantic
redis
openai
langchain
langchain_openai
dash_socketio
dash_mantine_components
dash_auth
uvicorn